+++ src/fides/api/service/connectors/sql_connector.py
@@ -4,6 +4,7 @@ from typing import Any, Dict, List, Optional, Type
 from urllib.parse import quote_plus
 
 import paramiko
+import pg8000
 import pymysql
 import sshtunnel  # type: ignore
 from aiohttp.client_exceptions import ClientResponseError
@@ -45,6 +46,9 @@ from fides.api.schemas.connection_configuration.connection_secrets_bigquery impo
 from fides.api.schemas.connection_configuration.connection_secrets_google_cloud_sql_mysql import (
     GoogleCloudSQLMySQLSchema,
 )
+from fides.api.schemas.connection_configuration.connection_secrets_google_cloud_sql_postgres import (
+    GoogleCloudSQLPostgresSchema,
+)
 from fides.api.schemas.connection_configuration.connection_secrets_mariadb import (
     MariaDBSchema,
 )
@@ -54,6 +58,7 @@ from fides.api.schemas.connection_configuration.connection_secrets_mysql import 
 from fides.api.service.connectors.base_connector import BaseConnector
 from fides.api.service.connectors.query_config import (
     BigQueryQueryConfig,
+    GoogleCloudSQLPostgresQueryConfig,
     MicrosoftSQLServerQueryConfig,
     RedshiftQueryConfig,
     SnowflakeQueryConfig,
@@ -625,3 +630,58 @@ class GoogleCloudSQLMySQLConnector(SQLConnector):
         """
         We need to override this method so it is not abstract anymore, and GoogleCloudSQLMySQLConnector is instantiable.
         """
+
+
+class GoogleCloudSQLPostgresConnector(SQLConnector):
+    """Connector specific to Google Cloud SQL for Postgres"""
+
+    secrets_schema = GoogleCloudSQLPostgresSchema
+
+    # Overrides SQLConnector.create_client
+    def create_client(self) -> Engine:
+        """Returns a SQLAlchemy Engine that can be used to interact with a database"""
+
+        config = self.secrets_schema(**self.configuration.secrets or {})
+
+        credentials = service_account.Credentials.from_service_account_info(
+            dict(config.keyfile_creds)
+        )
+
+        # initialize connector with the loaded credentials
+        connector = Connector(credentials=credentials)
+
+        def getconn() -> pg8000.dbapi.Connection:
+            conn: pg8000.dbapi.Connection = connector.connect(
+                config.instance_connection_name,
+                "pg8000",
+                user=config.db_iam_user,
+                db=config.dbname,
+                enable_iam_auth=True,
+            )
+            return conn
+
+        return create_engine("postgresql+pg8000://", creator=getconn)
+
+    @staticmethod
+    def cursor_result_to_rows(results: LegacyCursorResult) -> List[Row]:
+        """results to a list of dictionaries"""
+        return SQLConnector.default_cursor_result_to_rows(results)
+
+    def build_uri(self) -> None:
+        """
+        We need to override this method so it is not abstract anymore, and GoogleCloudSQLPostgresConnector is instantiable.
+        """
+
+    def set_schema(self, connection: Connection) -> None:
+        """Sets the schema for a postgres database if applicable"""
+        config = self.secrets_schema(**self.configuration.secrets or {})
+        if config.db_schema:
+            logger.info("Setting PostgreSQL search_path before retrieving data")
+            stmt = text("SELECT set_config('search_path', :search_path, false)")
+            stmt = stmt.bindparams(search_path=config.db_schema)
+            connection.execute(stmt)
+
+    # Overrides SQLConnector.query_config
+    def query_config(self, node: ExecutionNode) -> GoogleCloudSQLPostgresQueryConfig:
+        """Query wrapper corresponding to the input execution_node."""
+        return GoogleCloudSQLPostgresQueryConfig(node)